Remove support for the PC Mastercard plugin
The re-design of their website meant a complete rewrite of the PC
Mastercard plugin. I unfortunately did not succeed in figuring out how
the underlying API works and so this plugin is no longer functional. All
the work I did so far is available under the [pc-mastercard-redesign branch][1]
for anyone else to pick up.

[1]: https://github.com/marvinpinto/ledger-reconciler/commit/63401c4bb54d5c10ca142b47a91b7a0b1ad5d736